refactor(perf): split modular and namespaced TypeScript surfaces
